requirements.txt
## Blinky led_master python requirements
pymongo==3.6.0
tweepy==3.5.0
python-dateutil==2.6.1
colorama
coloredlogs
unidecode
textblob